Se realizo una refactorizacion al codigo, creando archivos js para cada componente
